fix infinite recursion loops for addSubItem, deleteSubItem and removeSubItem methods
The problem was the recursive call in the setParent method, the fix is to
remove controls done in this method and ensure the consistency in addSubItem,
deleteSubItem and removeSubItem methods.
The Item::~Item has also be changed to remove consistency checking.
This modification makes the delete and setParent methods unsafe to use in a
client code if the Item concerned has a not nullptr parent. In that case client
should call the appropriate method on the parent.