fix #5
Signature of all the find method changed : they return Item (and Item's
subclasses) lists to allow multiple Items with the same name in a hierarchy.
The Folder restriction is keeped : a Folder cannot contain two Items with the
same name. This restriction allow an easy implementation of the removeSubItem
method.
Documentation of the methods has been regenerated.