fix #7
The problem has been solved by adding a unregister loop on Item deletion.
An error has also been discovered : the unregister function raised an exception
if the unregistered Item was the last of the List (in that case the iterator
began equal to end() iterator).
This problem was also founded into Item::removeTag and was also fixed.