add TagManager::deleteTagsFromItem with a Tag vector as parameter
Improves performances if the caller even know the Tag list (otherwise it
should pass a string representing the Tags, which needed to be parsed and
extracted Tag refound).

Modification of the base Item deleteTags method with a call to this new method
(the Item know the Tag vector to unregister to).