add exception support in Item constructor and setName methods
Basic implementation is to ignore TaggerException and CoreException thrown
during the add/delete/updateTag methods. It may be overriden in subclasses if
an other behaviour is expected.

The main idea is to let the Item construct and setter finish if it cannot
update its tag list. The consistency error resulting is not worst enough to
justify an exception.