change TaggerException policy in create/delete methods from Item
The TaggerException is ignored in those methods, to allow the construction
of the Item if there is a problem during the Tag processing.

In fact, if an exception is thrown during a multiple registration process,
it would cause more consistency damages than other things : the resulting Item
would be part tagged. The actual solution tag completely the Item, and there
may be some Tag registering the wrong Item -this consistency error is due to
the programmer-