remove tag package and all dependancies
Actual Tag gesture is not interesting : global Tag list would grow
exponentially to the number of Item, and the gain for similar Item is not
good enough to be a viable solution.

An other point is the search engine that could search on specific fields, and
the Tag list would be processed on unnecessary tags.

Tag list may be a good solution but need to be detailed and applied to
specific fields to limit it cost.